+++ .github/workflows/scripts/create-github-release.sh
@@ -24,6 +24,8 @@ gh release create "$VERSION" \
   .genreleases/spec-kit-template-gemini-ps-"$VERSION".zip \
   .genreleases/spec-kit-template-cursor-agent-sh-"$VERSION".zip \
   .genreleases/spec-kit-template-cursor-agent-ps-"$VERSION".zip \
+  .genreleases/spec-kit-template-cline-sh-"$VERSION".zip \
+  .genreleases/spec-kit-template-cline-ps-"$VERSION".zip \
   .genreleases/spec-kit-template-opencode-sh-"$VERSION".zip \
   .genreleases/spec-kit-template-opencode-ps-"$VERSION".zip \
   .genreleases/spec-kit-template-qwen-sh-"$VERSION".zip \